Amendment 23 2070-01-23
Objectives

Discontinued reservation of seats for the Scheduled Tribes in Nagaland, both in the Lok Sabha and the State Legislative Assembly and stipulated that not more than one Anglo-Indian could be nominated by the Governor to any State Legislative Assembly. Extend reservation for SCs and STs and Anglo Indian members in the Lok Sabha and State Assemblies for another ten years, i.e. up to 1980.

Amendments

Amend articles 330, 332, 333 and 334.[27]

Closes #27

Signed-off-by: sushmit1 <sushmit718@gmail.com>